refactor: extrair OnboardingEditForm, OnboardingContent e ProjectDocs de ClientProfile.jsx
Resolve #51. Três componentes movidos para src/components/ClientProfile/:
- OnboardingEditForm.jsx — formulário de edição com upload de docs
- OnboardingContent.jsx — view de leitura do onboarding com toggle de edição
- ProjectDocs.jsx — chips de Raio-X/SLA com URLs assinadas

BUSINESS_LABELS e CONTRACT_MODEL_LABELS mantidos em ClientProfile.jsx pois
também são usados no header do componente principal.
